Support typescript return type of function type (#3069)
Fixes #3065

Note that I on purpose have set the domain the same as the content
range. Take this example:

`const foo: () => number;`

We already have a type scope with content range `() => number`. If we
set that as the domain for `number` it would probably be surprising if
you are inside the parameters in want to target the type.
@pokey Any thoughts on this?