Fix CONTRIBUTING.MDX formatting snafu introduced in PR 2989 (#3073)
From what I understand, the auto-formatting/prettifier plugin used by
Cursorless really didn't like how CONTRIBUTING.MDX needed to be
formatted to have the separation between vscode and vscodium specific
commands show properly, and as a result broke a bunch of the code
blocks, transforming them from
```bash
echo hello yes this is dog
```
to
`bash echo hello yes this is dog`
though ofc the commands were actual commands and not an echo example. I
forget if any other formatting was broken, it has been a while since I
actually finished the work on this and I just forgot to make a PR for
it, sorry (been busy with getting ready for the new semester, health
nonsense, and non-CS hobbies). This PR fixes the aforementioned issue
and excludes CONTRIBUTING.MDX from being processed by the
auto-formatting plugin, allowing it to contain all of the weird extended
markdown jank the plugin seems to really not like. If there's a more
elegant way to do this (changing the rules the plugin follows?) please
lmk and commit it if you have the time, this is very much not my usual
wheelhouse.

---------

Co-authored-by: Andreas Arvidsson <andreas.arvidsson87@gmail.com>