Update argument list domains and argument iteration domains in JS and TS (#3124)
I'm not a big fan of how complicated the function definitions in
javascript are becoming. Or how much code duplication there is.
The motivation being that if we really want argument list domain to be
the same as the function domain for all of these 25+ different patterns
we really need to combined them instead of trying to recreate them in a
separate argument list only definition.

@pokey Any input?